chore: Extract generic compose views to compose extensions
