+++ app/src/main/kotlin/org/eclipse/kuksa/testapp/databroker/viewmodel/ConnectionViewModel.kt
@@ -30,7 +30,7 @@ import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
-import org.eclipse.kuksa.testapp.model.ConnectionInfo
+import org.eclipse.kuksa.testapp.databroker.model.ConnectionInfo
 import org.eclipse.kuksa.testapp.preferences.ConnectionInfoRepository
 
 class ConnectionViewModel(application: Application) : AndroidViewModel(application) {
