+++ app/src/main/kotlin/org/eclipse/kuksa/testapp/extension/compose/HeadlineView.kt
@@ -1,0 +1,45 @@
+/*
+ *
+ *  * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *     http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *  *
+ *  * SPDX-License-Identifier: Apache-2.0
+ *  *
+ *
+ */
+
+package org.eclipse.kuksa.testapp.extension.compose
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun Headline(name: String, modifier: Modifier = Modifier, color: Color = Color.Black) {
+    Text(
+        text = name,
+        modifier = modifier
+            .fillMaxWidth()
+            .padding(horizontal = 15.dp),
+        textAlign = TextAlign.Center,
+        style = MaterialTheme.typography.titleLarge,
+        color = color,
+    )
+}
