--- app/src/main/kotlin/org/eclipse/kuksa/testapp/model/Certificate.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- *
- */
-
-package org.eclipse.kuksa.testapp.model
-
-import androidx.compose.runtime.Immutable
-import kotlinx.serialization.Serializable
-
-@Serializable
-@Immutable
-data class Certificate(
-    val name: String,
-    val extension: String,
-    val overrideAuthority: String = "",
-) {
-    var fullName = "$name.$extension"
-
-    companion object {
-        val DEFAULT = Certificate("CA", "pem")
-    }
-}
